Split components into structured directories
