+++ components/AlbumScreen/AlbumScreen.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Layout, Text } from '@ui-kitten/components';
+
+export default function AlbumScreen() {
+  return (
+    <Layout style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>相册管理</Text>
+    </Layout>
+  );
+}
