+++ components/CalenderScreen/MonthlyCalenderView/MonthlyCalenderView.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Layout, Text } from '@ui-kitten/components';
+
+export default function MonthlyCalenderView() {
+  return (
+    <Layout style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>日历月视图</Text>
+    </Layout>
+  );
+}
