+++ components/CalenderScreen/WeeklyCalenderView/WeeklyCalenderView.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Layout, Text } from '@ui-kitten/components';
+
+export default function WeeklyCalenderView() {
+  return (
+    <Layout style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>日历周视图</Text>
+    </Layout>
+  );
+}
