+++ components/ChoresScreen/ChoresScreen.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Layout, Text } from '@ui-kitten/components';
+
+export default function ChoresScreen() {
+  return (
+    <Layout style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>家务分配</Text>
+    </Layout>
+  );
+}
