+++ components/MyHomeScreen/MyHomeScreen.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Layout, Text } from '@ui-kitten/components';
+
+export default function MyHomeScreen() {
+  return (
+    <Layout style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+      <Text>我的家</Text>
+    </Layout>
+  );
+}
