+++ staticData.js
@@ -1,0 +1,26 @@
+export
+let eventsData = [
+  {
+    date: "7月27日",
+    events:
+      [
+        { name: "试镜", time: "18点 - 20点", location: "Stewart Center #320", memo: "Lorem ipsum dolor sit amet", organizer: "1" }
+      ]
+  },
+  {
+    date: "7月30日",
+    events:
+      [
+        { name: "约饭", time: "17点 - 19点", location: "家里", memo: "吃火锅" },
+        { name: "Watch Mulan on Disney+", time: "20点 - 22点", location: "家里", memo: "Lorem ipsum dolor sit amet", organizer: "2" }
+      ]
+  },
+  {
+    date: "8月1日",
+    events:
+      [
+        { name: "Watch Hamilton on Disney+", time: "18点 - 20点", location: "家里", memo: "Lorem ipsum dolor sit amet", organizer: "1" },
+        { name: "Rewatch Mulan on Disney+", time: "20点 - 22点", location: "家里", memo: "Lorem ipsum dolor sit amet", organizer: "2" }
+      ]
+  }
+];
